Replace git-clone with HTTP get of an archive
git-clone fetches the entire history of a repository. There is a GIT_SHALLOW
option in CMake's ExternalProject, but reportedly it can only be used to get
"named" commits (tags, branches).

ExternalProject provides an alternative way of fetching this, via URL combined
with github's ability to generate a zip archive of any arbitrary commit hash.

DOWNLOAD_EXTRACT_TIMESTAMP must be set to avoid warnings from CMake. Set to
True, the timestamps of the extracted files will match those in the archive.

I didn't add URL_HASH support mainly because this is more or less a PoC PR,
but if you choose to go this route, it should be added at some point. For now
I can provide md5 and sha256 hashes for your convenience.